test: migrate Footer, Gauge, Header tests
