+++ src/components/Gauge/__tests__/Gauge.spec.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import toJson from 'enzyme-to-json';
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom';
 import { Gauge } from '../Gauge';
 
 describe('Gauge', () => {
@@ -7,28 +8,34 @@ describe('Gauge', () => {
     mathRandomSpy.mockImplementation(() => 0.42);
 
     it('should render nothing if no children are provided', () => {
-        const wrapper = shallow(<Gauge />);
-        expect(wrapper.html()).toBeNull();
+        const view = render(<Gauge />);
+
+        expect(view.container).toMatchSnapshot();
     });
     it('should render children 0', () => {
-        const wrapper = shallow(<Gauge>0</Gauge>);
-        expect(wrapper.html()).not.toBeNull();
+        const view = render(<Gauge>0</Gauge>);
+
+        expect(view.container).toMatchSnapshot();
     });
     it('should render correctly', () => {
-        const wrapper = shallow(
+        const view = render(
             <Gauge percentage={33} note="Value" metric="Metric">
                 Some content
             </Gauge>
         );
-        expect(toJson(wrapper)).toMatchSnapshot();
+
+        expect(view.container).toMatchSnapshot();
+        expect(screen.queryByRole('alert')).not.toBeInTheDocument();
     });
 
     it('should render with loading state correctly', () => {
-        const wrapper = shallow(
+        const view = render(
             <Gauge metric="Metric" isContentLoading isProgressLoading>
                 Some content
             </Gauge>
         );
-        expect(wrapper.find('ContentPlaceholder').exists()).toBeTruthy();
+
+        expect(view.container).toMatchSnapshot();
+        expect(screen.getByRole('alert')).toBeInTheDocument();
     });
 });
